requirements.txt
aiosmtplib
aiogram
python-dotenv
